+++ R/bbox.R
@@ -21,17 +21,25 @@
 #' gbif_bbox2wkt(bbox=c(-125.0,38.4,-121.8,40.9))
 #'
 #' ## Or pass in each value separately
-#' gbif_bbox2wkt(minx=38.4, miny=-125.0, maxx=40.9, maxy=-121.8)
+#' gbif_bbox2wkt(minx=-125.0, miny=38.4, maxx=-121.8, maxy=40.9)
 #'
 #' # Convert a WKT object to a bounding box
-#' wkt <- "POLYGON((38.4 -125,40.9 -125,40.9 -121.8,38.4 -121.8,38.4 -125))"
+#' wkt <- "POLYGON((-125 38.4,-125 40.9,-121.8 40.9,-121.8 38.4,-125 38.4))"
 #' gbif_wkt2bbox(wkt)
 #' }
 gbif_bbox2wkt <- function(minx=NA, miny=NA, maxx=NA, maxy=NA, bbox=NULL){
   if (is.null(bbox)) bbox <- c(minx, miny, maxx, maxy)
   stopifnot(noNA(bbox)) #check for NAs
   stopifnot(is.numeric(as.numeric(bbox))) #check for numeric-ness
-  wicket::bounding_wkt(values = bbox)
+  # wicket::bounding_wkt(values = bbox)
+  bbox_template <- 'POLYGON((%s %s,%s %s,%s %s,%s %s,%s %s))'
+  sprintf(bbox_template, 
+    bbox[1], bbox[2],
+    bbox[3], bbox[2],
+    bbox[3], bbox[4],
+    bbox[1], bbox[4],
+    bbox[1], bbox[2]
+  )
 }
 
 #' @export