+++ R/occ_issues.r
@@ -1,19 +1,23 @@
-#' Parse and examine further GBIF issues on a dataset.
+#' Parse and examine further GBIF issues on a dataset
 #'
 #' @export
 #'
-#' @param .data Output from a call to [occ_search()], but only if
-#' `return="all"`, or `return="data"`, otherwise function stops
-#' with error
+#' @param .data Output from a call to [occ_search()], [occ_data()], or 
+#' [occ_download_import()], but only if `return="all"`, or `return="data"`, 
+#' otherwise function stops with error. The data from `occ_download_import`
+#' is just a regular data.frame so you can pass in a data.frame to this
+#' function, but if it doesn't have certain columns it will fail.
 #' @param ... Named parameters to only get back (e.g., cdround), or to
 #' remove (e.g. -cdround).
 #' @param mutate (character) One of:
-#' \itemize{
-#'  \item split Split issues into new columns.
-#'  \item split_expand Split into new columns, and expand issue names.
-#'  \item expand Expand issue abbreviated codes into descriptive names.
-#' }
-#' For split and split_expand, values in cells become y ("yes") or n ("no").
+#' 
+#' - `split` Split issues into new columns
+#' - `split_expand` Split into new columns, and expand issue names
+#' - `expand` Expand issue abbreviated codes into descriptive names. 
+#' for downloads datasets, this is not super useful since the 
+#' issues come to you as expanded already.
+#' 
+#' For split and split_expand, values in cells become y ("yes") or n ("no")
 #'
 #' @references
 #' <http://gbif.github.io/gbif-api/apidocs/org/gbif/api/vocabulary/OccurrenceIssue.html>
@@ -30,6 +34,10 @@
 #' returned from a call to [occ_search()]. Maybe in a future version
 #' we will remove the associated records from the `hierarchy` and `media`
 #' elements as they are remove from the `data` element.
+#' 
+#' You'll notice that we sort columns to make it easier to glimpse the important
+#' parts of your data, namely taxonomic name, taxon key, latitude and longitude, 
+#' and the issues. The columns are unchanged otherwise.
 #'
 #' @examples \dontrun{
 #' ## what do issues mean, can print whole table, or search for matches
@@ -38,14 +46,20 @@
 #'
 #' # compare out data to after occ_issues use
 #' (out <- occ_search(limit=100))
-#' out %>% occ_issues(cudc)
+#' out %>% occ_issues(cdround)
+#' 
+#' # occ_data
+#' (out <- occ_data(limit=100))
+#' out %>% occ_issues(cdround)
 #'
 #' # Parsing output by issue
-#' (res <- occ_search(geometry='POLYGON((30.1 10.1, 10 20, 20 40, 40 40, 30.1 10.1))', limit = 50))
+#' (res <- occ_data(
+#'   geometry='POLYGON((30.1 10.1, 10 20, 20 40, 40 40, 30.1 10.1))', 
+#'   limit = 600))
 #'
 #' ## or parse issues in various ways
-#' ### inlude only rows with gass84 issue
-#' gg <- res %>% occ_issues(gass84)
+#' ### inlude only rows with cdround issue
+#' gg <- res %>% occ_issues(cdround)
 #' NROW(res$data)
 #' NROW(gg$data)
 #' head(res$data)[,c(1:5)]
@@ -66,19 +80,51 @@
 #' res %>% occ_issues(mutate = "split_expand")
 #'
 #' ### split, expand, and remove an issue class
-#' res %>% occ_issues(-cudc, mutate = "split_expand")
+#' res %>% occ_issues(-cdround, mutate = "split_expand")
 #'
 #' ## Or you can use occ_issues without %>%
-#' occ_issues(res, -cudc, mutate = "split_expand")
+#' occ_issues(res, -cdround, mutate = "split_expand")
+#' 
+#' # from GBIF downloaded data via occ_download and friends
+#' res <- occ_download_get(key="0000066-140928181241064", overwrite=TRUE)
+#' x <- occ_download_import(res)
+#' occ_issues(x, -txmathi)
+#' occ_issues(x, txmathi)
+#' occ_issues(x, gass84)
+#' occ_issues(x, zerocd)
+#' occ_issues(x, gass84, txmathi)
+#' occ_issues(x, mutate = "split")
+#' occ_issues(x, -gass84, mutate = "split")
+#' occ_issues(x, mutate = "expand")
+#' occ_issues(x, mutate = "split_expand")
 #' }
 
 occ_issues <- function(.data, ..., mutate = NULL) {
 
-  stopifnot(xor(inherits(.data, "gbif"), inherits(.data, "gbif_data")))
+  # stopifnot(xor(inherits(.data, "gbif"), inherits(.data, "gbif_data")))
+  assert(.data, c("gbif", "gbif_data", "data.frame", "tbl_df"))
+
   if ("data" %in% names(.data)) {
     tmp <- .data$data
   } else {
     tmp <- .data
+  }
+
+  # handle downloads data 
+  is_dload <- FALSE
+  if (
+    c("issue", "gbifID", "accessRights", "accrualMethod") %in% names(tmp) &&
+    !"issues" %in% names(tmp)
+  ) {
+    is_dload <- TRUE
+
+    # convert long issue names to short ones
+    issstr <- tmp[names(tmp) %in% "issue"][[1]]
+    tmp$issues <- unlist(lapply(issstr, function(z) {
+      if (identical(z, "")) return (character(1))
+      paste(gbifissues[ grepl(gsub(";", "|", z), gbifissues$issue), "code" ], 
+        collapse = ",")
+    }))
   }
 
   if (!length(dots(...)) == 0) {
@@ -93,10 +139,10 @@ occ_issues <- function(.data, ..., mutate = NULL) {
 
   if (!is.null(mutate)) {
     if (mutate == 'split') {
-      tmp <- split_iss(tmp)
+      tmp <- split_iss(tmp, is_dload)
     } else if (mutate == 'split_expand') {
       tmp <- mutate_iss(tmp)
-      tmp <- split_iss(tmp)
+      tmp <- split_iss(tmp, is_dload)
     } else if (mutate == 'expand') {
       tmp <- mutate_iss(tmp)
     }
@@ -110,14 +156,14 @@ occ_issues <- function(.data, ..., mutate = NULL) {
   }
 }
 
-mutate_iss <- function(w){
+mutate_iss <- function(w) {
   w$issues <- sapply(strsplit(w$issues, split = ","), function(z) {
     paste(gbifissues[ gbifissues$code %in% z, "issue" ], collapse = ",")
   })
   return( w )
 }
 
-split_iss <- function(m){
+split_iss <- function(m, is_dload) {
   unq <- unique(unlist(strsplit(m$issues, split = ",")))
   df <- data.table::setDF(
     data.table::rbindlist(
@@ -129,8 +175,11 @@ split_iss <- function(m){
   )
   names(df) <- unq
   m$issues <- NULL
-  first <- c('name','key','decimalLatitude','decimalLongitude')
-  tibble::as_data_frame(data.frame(m[, first], df, m[, !names(m) %in% first], stringsAsFactors = FALSE))
+  first_search <- c('name','key','decimalLatitude','decimalLongitude')
+  first_dload <- c('scientificName','taxonKey','decimalLatitude','decimalLongitude')
+  first <- if (is_dload) first_dload else first_search
+  tibble::as_data_frame(data.frame(m[, first], df, m[, !names(m) %in% first], 
+    stringsAsFactors = FALSE))
 }
 
 parse_input <- function(...) {