+++ tests/testthat/test-map_fetch.R
@@ -3,7 +3,7 @@ context("map_fetch")
 test_that("map_fetch", {
   skip_on_cran()
 
-  x <- map_fetch(search = "taxonKey", id = 3118771, year = 2010)
+  x <- map_fetch(taxonKey = 3118771, year = 2010)
   expect_is(x, "RasterLayer")
 })
 
@@ -11,7 +11,11 @@ test_that("map_fetch", {
 test_that("map_fetch fails well", {
   skip_on_cran()
 
-  expect_error(map_fetch(source = "stuff"), "'arg' should be one of")
+  expect_error(map_fetch(source = "stuff"), "is not TRUE")
   expect_error(map_fetch(source = 5), "source must be of class character")
-  expect_error(map_fetch(format = 'bears'), "'arg' should be one of")
+  expect_error(map_fetch(format = 'bears'), "is not TRUE")
+
+  # search and id have been removed
+  expect_error(map_fetch(search = 'bears'), "have been removed")
+  expect_error(map_fetch(id = 3), "have been removed")
 })